Enhance Sample and provide XsuaaExtension Make SecurityTestExtension abstract Rename SecurityTestConfiguration to SecurityTestContext
