+++ java-security-test/src/main/java/com/sap/cloud/security/test/extension/XsuaaExtension.java
@@ -1,0 +1,11 @@
+package com.sap.cloud.security.test.extension;
+
+import com.sap.cloud.security.config.Service;
+import com.sap.cloud.security.test.SecurityTest;
+
+public class XsuaaExtension extends SecurityTestExtension {
+
+    public XsuaaExtension() {
+        super(new SecurityTest(Service.XSUAA));
+    }
+}
